Add layout to webpage
Added layout including a header bar, modal and main content of the
webpage.

Header bar contains title of the webpage.

Modal is a form where the user enters details of new book they want to
add to the library.

Main content includes, an 'add book' button that brings up the modal
form. Also a table display the books in the library.

Each table row displaying a book has two buttons, one to change the
read/unread status and another to remove the book from the library.